install: clang-11.x compatible PGO instructions
LLVM/Clang 11.x throws errors for _files_ without any profiling data:

  error: ../../async.c: Function control flow change detected (hash mismatch) async_write Hash = 72057641800614222 [-Werror,-Wbackend-plugin]

This caused multiple issues with the current PGO instructions:

* The `pgo` binary failed to link with `meson configure -Db_pgo=use`
  when doing a full PGO build (since it isn't used in this case);
* `footclient` fails to link for the same reason.
* `foot` fails to link for the same reason in **partial** PGO builds.

The solution is to make sure all binaries that are built in the final
phase (`-Db_pgo=use`) have been executed in the _generate_ phase.

Doing this also means we can drop `-Wno-missing-profile` and friends.

Also add `--sixel` to the `generate-alt-random-writes` command line in
the description for a full PGO build.

Closes #418